2.7.1 : Added img and thread packages. Redesigned cv package and created several classes like Hardware, Metric, TimeClock, Calc, Drawing, etc. Log was improved and fixed some bugs in previous version. Updated dependencies.

+++ modules/web/src/test/groovy/com/intellisrc/web/WebServiceTest.groovy
@@ -53,7 +53,9 @@ class WebServiceTest extends Specification {
         then:
             assert !web.isRunning()
         cleanup:
-            Log.logFile.delete()
+            if(Log.logFile && Log.logFile.exists()) {
+                Log.logFile.delete()
+            }
     }
 
     def "Testing auto cache"() {
